Function procout::procout

pub fn procout(
    code_block: &TokenStream,
    module_ident: Option<Ident>,
    output_path: Option<&str>,
)

Handle printing code to a file

code_block This is the code that should be printed (the TokenStream output of the macro being debugged)

module_ident This is the optional name of the module generated by the macro.

output_path This is the directory to write the file to.